Refactor indexer_ui -> ui
